Create builder helpers for Menus
This allows setting up menus via a builder-style API rather than by
passing a table to the constructor.

This should be a more manageable pattern for setting up things like menu
breadcrumbs, etc, where we want to manage back and forth between both
the menu and the binding/action we're setting up.